Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-values/animations/calc-interpolation-expected.txt


FAIL CSS Transitions: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (-0.25) should be [-10px] assert_equals: expected "- 10px " but got "0px "
PASS CSS Transitions: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0) should be [0px]
PASS CSS Transitions: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.25) should be [10px]
PASS CSS Transitions: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.5) should be [20px]
PASS CSS Transitions: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.75) should be [30px]
PASS CSS Transitions: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1) should be [40px]
FAIL CSS Transitions: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1.25) should be [50px] assert_equals: expected "50px " but got "40px "
FAIL CSS Transitions with transition: all: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (-0.25) should be [-10px] assert_equals: expected "- 10px " but got "0px "
PASS CSS Transitions with transition: all: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0) should be [0px]
PASS CSS Transitions with transition: all: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.25) should be [10px]
PASS CSS Transitions with transition: all: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.5) should be [20px]
PASS CSS Transitions with transition: all: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.75) should be [30px]
PASS CSS Transitions with transition: all: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1) should be [40px]
FAIL CSS Transitions with transition: all: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1.25) should be [50px] assert_equals: expected "50px " but got "40px "
FAIL CSS Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (-0.25) should be [-10px] assert_equals: expected "- 10px " but got "0px "
PASS CSS Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0) should be [0px]
PASS CSS Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.25) should be [10px]
PASS CSS Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.5) should be [20px]
PASS CSS Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.75) should be [30px]
PASS CSS Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1) should be [40px]
FAIL CSS Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1.25) should be [50px] assert_equals: expected "50px " but got "40px "
FAIL Web Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (-0.25) should be [-10px] assert_equals: expected "- 10px " but got "0px "
PASS Web Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0) should be [0px]
PASS Web Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.25) should be [10px]
PASS Web Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.5) should be [20px]
PASS Web Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.75) should be [30px]
PASS Web Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1) should be [40px]
FAIL Web Animations: property <left> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1.25) should be [50px] assert_equals: expected "50px " but got "40px "
FAIL CSS Transitions: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (-0.25) should be [calc(((50% - 25px) * 1.25) + ((100% - 10px) * -0.25))] assert_equals: expected "calc ( ( 1.25 * ( 50 % - 25px ) ) + ( - 0.25 * ( 100 % - 10px ) ) ) " but got "calc ( 50 % - 25px ) "
PASS CSS Transitions: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0) should be [calc(50% - 25px)]
FAIL CSS Transitions: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.25) should be [calc(((50% - 25px) * 0.75) + ((100% - 10px) * 0.25))] assert_equals: expected "calc ( ( 0.75 * ( 50 % - 25px ) ) + ( 0.25 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.75 ) + ( ( 100 % - 10px ) * 0.25 ) ) "
FAIL CSS Transitions: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.5) should be [calc(((50% - 25px) * 0.5) + ((100% - 10px) * 0.5))] assert_equals: expected "calc ( ( 0.5 * ( 50 % - 25px ) ) + ( 0.5 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.5 ) + ( ( 100 % - 10px ) * 0.5 ) ) "
FAIL CSS Transitions: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.75) should be [calc(((50% - 25px) * 0.25) + ((100% - 10px) * 0.75))] assert_equals: expected "calc ( ( 0.25 * ( 50 % - 25px ) ) + ( 0.75 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.25 ) + ( ( 100 % - 10px ) * 0.75 ) ) "
PASS CSS Transitions: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1) should be [calc(100% - 10px)]
FAIL CSS Transitions: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1.25) should be [calc(((50% - 25px) * -0.25) + ((100% - 10px) * 1.25))] assert_equals: expected "calc ( ( - 0.25 * ( 50 % - 25px ) ) + ( 1.25 * ( 100 % - 10px ) ) ) " but got "calc ( 100 % - 10px ) "
FAIL CSS Transitions with transition: all: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (-0.25) should be [calc(((50% - 25px) * 1.25) + ((100% - 10px) * -0.25))] assert_equals: expected "calc ( ( 1.25 * ( 50 % - 25px ) ) + ( - 0.25 * ( 100 % - 10px ) ) ) " but got "calc ( 50 % - 25px ) "
PASS CSS Transitions with transition: all: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0) should be [calc(50% - 25px)]
FAIL CSS Transitions with transition: all: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.25) should be [calc(((50% - 25px) * 0.75) + ((100% - 10px) * 0.25))] assert_equals: expected "calc ( ( 0.75 * ( 50 % - 25px ) ) + ( 0.25 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.75 ) + ( ( 100 % - 10px ) * 0.25 ) ) "
FAIL CSS Transitions with transition: all: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.5) should be [calc(((50% - 25px) * 0.5) + ((100% - 10px) * 0.5))] assert_equals: expected "calc ( ( 0.5 * ( 50 % - 25px ) ) + ( 0.5 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.5 ) + ( ( 100 % - 10px ) * 0.5 ) ) "
FAIL CSS Transitions with transition: all: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.75) should be [calc(((50% - 25px) * 0.25) + ((100% - 10px) * 0.75))] assert_equals: expected "calc ( ( 0.25 * ( 50 % - 25px ) ) + ( 0.75 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.25 ) + ( ( 100 % - 10px ) * 0.75 ) ) "
PASS CSS Transitions with transition: all: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1) should be [calc(100% - 10px)]
FAIL CSS Transitions with transition: all: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1.25) should be [calc(((50% - 25px) * -0.25) + ((100% - 10px) * 1.25))] assert_equals: expected "calc ( ( - 0.25 * ( 50 % - 25px ) ) + ( 1.25 * ( 100 % - 10px ) ) ) " but got "calc ( 100 % - 10px ) "
FAIL CSS Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (-0.25) should be [calc(((50% - 25px) * 1.25) + ((100% - 10px) * -0.25))] assert_equals: expected "calc ( ( 1.25 * ( 50 % - 25px ) ) + ( - 0.25 * ( 100 % - 10px ) ) ) " but got "calc ( 50 % - 25px ) "
PASS CSS Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0) should be [calc(50% - 25px)]
FAIL CSS Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.25) should be [calc(((50% - 25px) * 0.75) + ((100% - 10px) * 0.25))] assert_equals: expected "calc ( ( 0.75 * ( 50 % - 25px ) ) + ( 0.25 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.75 ) + ( ( 100 % - 10px ) * 0.25 ) ) "
FAIL CSS Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.5) should be [calc(((50% - 25px) * 0.5) + ((100% - 10px) * 0.5))] assert_equals: expected "calc ( ( 0.5 * ( 50 % - 25px ) ) + ( 0.5 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.5 ) + ( ( 100 % - 10px ) * 0.5 ) ) "
FAIL CSS Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.75) should be [calc(((50% - 25px) * 0.25) + ((100% - 10px) * 0.75))] assert_equals: expected "calc ( ( 0.25 * ( 50 % - 25px ) ) + ( 0.75 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.25 ) + ( ( 100 % - 10px ) * 0.75 ) ) "
PASS CSS Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1) should be [calc(100% - 10px)]
FAIL CSS Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1.25) should be [calc(((50% - 25px) * -0.25) + ((100% - 10px) * 1.25))] assert_equals: expected "calc ( ( - 0.25 * ( 50 % - 25px ) ) + ( 1.25 * ( 100 % - 10px ) ) ) " but got "calc ( 100 % - 10px ) "
FAIL Web Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (-0.25) should be [calc(((50% - 25px) * 1.25) + ((100% - 10px) * -0.25))] assert_equals: expected "calc ( ( 1.25 * ( 50 % - 25px ) ) + ( - 0.25 * ( 100 % - 10px ) ) ) " but got "calc ( 50 % - 25px ) "
PASS Web Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0) should be [calc(50% - 25px)]
FAIL Web Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.25) should be [calc(((50% - 25px) * 0.75) + ((100% - 10px) * 0.25))] assert_equals: expected "calc ( ( 0.75 * ( 50 % - 25px ) ) + ( 0.25 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.75 ) + ( ( 100 % - 10px ) * 0.25 ) ) "
FAIL Web Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.5) should be [calc(((50% - 25px) * 0.5) + ((100% - 10px) * 0.5))] assert_equals: expected "calc ( ( 0.5 * ( 50 % - 25px ) ) + ( 0.5 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.5 ) + ( ( 100 % - 10px ) * 0.5 ) ) "
FAIL Web Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (0.75) should be [calc(((50% - 25px) * 0.25) + ((100% - 10px) * 0.75))] assert_equals: expected "calc ( ( 0.25 * ( 50 % - 25px ) ) + ( 0.75 * ( 100 % - 10px ) ) ) " but got "calc ( ( ( 50 % - 25px ) * 0.25 ) + ( ( 100 % - 10px ) * 0.75 ) ) "
PASS Web Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1) should be [calc(100% - 10px)]
FAIL Web Animations: property <text-indent> from [calc(50% - 25px)] to [calc(100% - 10px)] at (1.25) should be [calc(((50% - 25px) * -0.25) + ((100% - 10px) * 1.25))] assert_equals: expected "calc ( ( - 0.25 * ( 50 % - 25px ) ) + ( 1.25 * ( 100 % - 10px ) ) ) " but got "calc ( 100 % - 10px ) "
PASS CSS Transitions: property <text-indent> from [0em] to [100px] at (-0.25) should be [-25px]
PASS CSS Transitions: property <text-indent> from [0em] to [100px] at (0) should be [0em]
PASS CSS Transitions: property <text-indent> from [0em] to [100px] at (0.25) should be [25px]
PASS CSS Transitions: property <text-indent> from [0em] to [100px] at (0.5) should be [50px]
PASS CSS Transitions: property <text-indent> from [0em] to [100px] at (0.75) should be [75px]
PASS CSS Transitions: property <text-indent> from [0em] to [100px] at (1) should be [100px]
PASS CSS Transitions: property <text-indent> from [0em] to [100px] at (1.25) should be [125px]
PASS CSS Transitions with transition: all: property <text-indent> from [0em] to [100px] at (-0.25) should be [-25px]
PASS CSS Transitions with transition: all: property <text-indent> from [0em] to [100px] at (0) should be [0em]
PASS CSS Transitions with transition: all: property <text-indent> from [0em] to [100px] at (0.25) should be [25px]
PASS CSS Transitions with transition: all: property <text-indent> from [0em] to [100px] at (0.5) should be [50px]
PASS CSS Transitions with transition: all: property <text-indent> from [0em] to [100px] at (0.75) should be [75px]
PASS CSS Transitions with transition: all: property <text-indent> from [0em] to [100px] at (1) should be [100px]
PASS CSS Transitions with transition: all: property <text-indent> from [0em] to [100px] at (1.25) should be [125px]
PASS CSS Animations: property <text-indent> from [0em] to [100px] at (-0.25) should be [-25px]
PASS CSS Animations: property <text-indent> from [0em] to [100px] at (0) should be [0em]
PASS CSS Animations: property <text-indent> from [0em] to [100px] at (0.25) should be [25px]
PASS CSS Animations: property <text-indent> from [0em] to [100px] at (0.5) should be [50px]
PASS CSS Animations: property <text-indent> from [0em] to [100px] at (0.75) should be [75px]
PASS CSS Animations: property <text-indent> from [0em] to [100px] at (1) should be [100px]
PASS CSS Animations: property <text-indent> from [0em] to [100px] at (1.25) should be [125px]
PASS Web Animations: property <text-indent> from [0em] to [100px] at (-0.25) should be [-25px]
PASS Web Animations: property <text-indent> from [0em] to [100px] at (0) should be [0em]
PASS Web Animations: property <text-indent> from [0em] to [100px] at (0.25) should be [25px]
PASS Web Animations: property <text-indent> from [0em] to [100px] at (0.5) should be [50px]
PASS Web Animations: property <text-indent> from [0em] to [100px] at (0.75) should be [75px]
PASS Web Animations: property <text-indent> from [0em] to [100px] at (1) should be [100px]
PASS Web Animations: property <text-indent> from [0em] to [100px] at (1.25) should be [125px]
FAIL CSS Transitions: property <text-indent> from [0%] to [100px] at (-0.25) should be [calc(0% + -25px)] assert_equals: expected "calc ( 0 % - 25px ) " but got "- 25px "
FAIL CSS Transitions: property <text-indent> from [0%] to [100px] at (0) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions: property <text-indent> from [0%] to [100px] at (0.25) should be [calc(0% + 25px)] assert_equals: expected "calc ( 0 % + 25px ) " but got "25px "
FAIL CSS Transitions: property <text-indent> from [0%] to [100px] at (0.5) should be [calc(0% + 50px)] assert_equals: expected "calc ( 0 % + 50px ) " but got "50px "
FAIL CSS Transitions: property <text-indent> from [0%] to [100px] at (0.75) should be [calc(0% + 75px)] assert_equals: expected "calc ( 0 % + 75px ) " but got "75px "
FAIL CSS Transitions: property <text-indent> from [0%] to [100px] at (1) should be [calc(0% + 100px)] assert_equals: expected "calc ( 0 % + 100px ) " but got "100px "
FAIL CSS Transitions: property <text-indent> from [0%] to [100px] at (1.25) should be [calc(0% + 125px)] assert_equals: expected "calc ( 0 % + 125px ) " but got "125px "
FAIL CSS Transitions with transition: all: property <text-indent> from [0%] to [100px] at (-0.25) should be [calc(0% + -25px)] assert_equals: expected "calc ( 0 % - 25px ) " but got "- 25px "
FAIL CSS Transitions with transition: all: property <text-indent> from [0%] to [100px] at (0) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Transitions with transition: all: property <text-indent> from [0%] to [100px] at (0.25) should be [calc(0% + 25px)] assert_equals: expected "calc ( 0 % + 25px ) " but got "25px "
FAIL CSS Transitions with transition: all: property <text-indent> from [0%] to [100px] at (0.5) should be [calc(0% + 50px)] assert_equals: expected "calc ( 0 % + 50px ) " but got "50px "
FAIL CSS Transitions with transition: all: property <text-indent> from [0%] to [100px] at (0.75) should be [calc(0% + 75px)] assert_equals: expected "calc ( 0 % + 75px ) " but got "75px "
FAIL CSS Transitions with transition: all: property <text-indent> from [0%] to [100px] at (1) should be [calc(0% + 100px)] assert_equals: expected "calc ( 0 % + 100px ) " but got "100px "
FAIL CSS Transitions with transition: all: property <text-indent> from [0%] to [100px] at (1.25) should be [calc(0% + 125px)] assert_equals: expected "calc ( 0 % + 125px ) " but got "125px "
FAIL CSS Animations: property <text-indent> from [0%] to [100px] at (-0.25) should be [calc(0% + -25px)] assert_equals: expected "calc ( 0 % - 25px ) " but got "- 25px "
FAIL CSS Animations: property <text-indent> from [0%] to [100px] at (0) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL CSS Animations: property <text-indent> from [0%] to [100px] at (0.25) should be [calc(0% + 25px)] assert_equals: expected "calc ( 0 % + 25px ) " but got "25px "
FAIL CSS Animations: property <text-indent> from [0%] to [100px] at (0.5) should be [calc(0% + 50px)] assert_equals: expected "calc ( 0 % + 50px ) " but got "50px "
FAIL CSS Animations: property <text-indent> from [0%] to [100px] at (0.75) should be [calc(0% + 75px)] assert_equals: expected "calc ( 0 % + 75px ) " but got "75px "
FAIL CSS Animations: property <text-indent> from [0%] to [100px] at (1) should be [calc(0% + 100px)] assert_equals: expected "calc ( 0 % + 100px ) " but got "100px "
FAIL CSS Animations: property <text-indent> from [0%] to [100px] at (1.25) should be [calc(0% + 125px)] assert_equals: expected "calc ( 0 % + 125px ) " but got "125px "
FAIL Web Animations: property <text-indent> from [0%] to [100px] at (-0.25) should be [calc(0% + -25px)] assert_equals: expected "calc ( 0 % - 25px ) " but got "- 25px "
FAIL Web Animations: property <text-indent> from [0%] to [100px] at (0) should be [0%] assert_equals: expected "0 % " but got "0px "
FAIL Web Animations: property <text-indent> from [0%] to [100px] at (0.25) should be [calc(0% + 25px)] assert_equals: expected "calc ( 0 % + 25px ) " but got "25px "
FAIL Web Animations: property <text-indent> from [0%] to [100px] at (0.5) should be [calc(0% + 50px)] assert_equals: expected "calc ( 0 % + 50px ) " but got "50px "
FAIL Web Animations: property <text-indent> from [0%] to [100px] at (0.75) should be [calc(0% + 75px)] assert_equals: expected "calc ( 0 % + 75px ) " but got "75px "
FAIL Web Animations: property <text-indent> from [0%] to [100px] at (1) should be [calc(0% + 100px)] assert_equals: expected "calc ( 0 % + 100px ) " but got "100px "
FAIL Web Animations: property <text-indent> from [0%] to [100px] at (1.25) should be [calc(0% + 125px)] assert_equals: expected "calc ( 0 % + 125px ) " but got "125px "
